feat: rename "ModelNameCommand" to "PerformanceCommand"
