+++ src/MultiTekla.Plugins/Headless/Performance/PerformancePlugin.cs
@@ -1,0 +1,7 @@
+namespace MultiTekla.Plugins.Headless.Performance;
+
+public class PerformancePlugin : PluginBase<bool>
+{
+    protected override bool Run()
+        => new TSM.Model().GetConnectionStatus();
+}
