+++ src/MultiTekla.Plugins/Model/ModelCommand.cs
@@ -1,6 +1,6 @@
 using CliFx;
 
-namespace MultiTekla.Plugins;
+namespace MultiTekla.Plugins.Model;
 
 [Command("model", Description = "Model related commands")]
 public class ModelCommand : ICommand
